Move loadSigningKeys to separated utils file
