+++ src/ui/states/approveTransaction/ApproveTransactionState.tsx
@@ -11,6 +11,7 @@ import { withAccessControl } from '../../components/auth/withAccessControl'
 import type { OutletContext } from '../../types/OutletContext'
 import LoadingScreen from '../../components/layout/LoadingScreen'
 import { PublicKeySelect, PublicKeyInfo } from '../../components/keys'
+import { loadSigningKeys as getSigningKeys } from '../../../utils'
 
 function ApproveTransactionState (): React.JSX.Element {
   const navigate = useNavigate()
@@ -93,33 +94,7 @@ function ApproveTransactionState (): React.JSX.Element {
       const wallet = allWallets.find(w => w.walletId === currentWallet && w.network === currentNetwork)
       if (wallet == null) throw new Error('Wallet not found')
 
-      const identityPublicKeys = await sdk.identities.getIdentityPublicKeys(currentIdentity)
-      const availableKeyIds = await extensionAPI.getAvailableKeyPairs(currentIdentity)
-
-      // Filter identity public keys to only show those that are available
-      const availablePublicKeys = identityPublicKeys.filter((key: any) => {
-        const keyId = key?.keyId ?? key?.getId?.() ?? null
-        return keyId != null && availableKeyIds.includes(keyId)
-      })
-
-      const keys: PublicKeyInfo[] = availablePublicKeys.map((key: any) => {
-        const keyId = key?.keyId ?? key?.getId?.() ?? null
-        const purpose = String(key?.purpose ?? 'UNKNOWN')
-        const security = String(key?.securityLevel ?? 'UNKNOWN')
-        let hash = ''
-        try {
-          hash = typeof key?.getPublicKeyHash === 'function' ? key.getPublicKeyHash() : ''
-        } catch {}
-
-        return {
-          keyId: keyId ?? 0,
-          securityLevel: security,
-          purpose,
-          hash
-        }
-      })
-
-      return keys
+      return await getSigningKeys(sdk, extensionAPI, currentIdentity)
     })
       .catch(e => console.log('loadSigningKeys error', e))
   }, [currentWallet, currentNetwork, currentIdentity])
