+++ src/utils/index.ts
@@ -9,6 +9,7 @@ import { decrypt, PrivateKey } from 'eciesjs'
 import { KeypairRepository } from '../content-script/repository/KeypairRepository'
 
 export { formatBigNumber }
+export { loadSigningKeys } from './signingKeys'
 
 export const hexToBytes = (hex: string): Uint8Array => {
   return Uint8Array.from((hex.match(/.{1,2}/g) ?? []).map((byte) => parseInt(byte, 16)))
